Create setup.sh from base-image
Right now the user who needs to add packages and otherwise modify the basic environment needs to remember to change it in two places. If he forgets it can be especially tough to debug, since it seems to work locally. Constructing the setup _from_ the Dockerfile maintains that single point of truth.